Refactor home view into daily quote.
