+++ daily_quote/views.py
@@ -1,7 +1,7 @@
 from django.http import Http404
 from django.shortcuts import render
 
-from daily_quote.models import Profile
+from daily_quote.models import Profile, Quote
 
 
 def user_profile(request, username):
@@ -27,3 +27,17 @@ def user_profile(request, username):
         raise Http404("Profile Does Not Exist")
 
     return render(request, 'daily_quote/profile.html', context)
+
+
+def home(request):
+    if request.user.is_authenticated:
+        quote = request.user.profile.recommend().quote
+    else:
+        quote = Quote.random_quote()
+
+    context = {
+        'current_route': 'home',
+        'quote': quote
+    }
+
+    return render(request, 'daily_quote/home.html', context)
