+++ templates/daily_quote/home.html
